Further updates to ProgressBar
This adds an update() method that should make it easier to use for
providing a running update in addition to just a percentage / busy
indicator.

Also ported tckgen & tckedit to use this method.

+++ src/dwi/tractography/file_base.h
@@ -127,7 +127,7 @@ namespace MR
           }
 
 
-          size_t count, total_count;
+          uint64_t count, total_count;
 
 
         protected:
